Added Dofus portal command
The portal command retrieves the portal's position and uses left and send a formatted embedded msg to the channel. The data is take by web scraping the dofus-portals.fr website with a python script and running this script through node.js.
Right now the Dofus server this applies for is hardcoded